Switched money calculations from BigInteger to BigDecimal Improved BasePromotion.apply Started work on README.md

+++ src/main/java/org/duncati/promotionengine/BundlePromotion.java
@@ -1,6 +1,6 @@
 package org.duncati.promotionengine;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 /**
  * This promotion defines a static price for a given set of items, for example a single item A might cost 50
@@ -10,19 +10,19 @@ public class BundlePromotion extends BasePromotion {
 
     // simple flat fee promotion offering lower price for a set (bundle) of items
 
-    private final BigInteger price;
+    private final BigDecimal price;
 
     /**
      * Creates a bundle promotion for the given items at the given price
      * @param items the items in this promotion
      * @param price the price for this promotion
      */
-    public BundlePromotion(Items items, BigInteger price) {
+    public BundlePromotion(Items items, BigDecimal price) {
         super(items);
         this.price=price;
     }
 
-    public BigInteger getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
